% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmixLitMiner-package.R
\docType{package}
\name{OmixLitMiner}
\alias{OmixLitMiner}
\alias{OmixLitMiner-package}
\title{OmixLitMiner: A package for automating literatue retrieval from PubMed based on UniProt identifiers and user-defined keyword search..}
\source{
Steffen P, Wu J, Hariharan S, Molloy MP, Schluter H
Bowel Cancer and Biomarker Research, Kolling Institute, The University of Sydney, Australia
}
\description{
The OmixLitMiner package provides functions:
omixLitMiner
}
