% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PubMedMiner.R
\name{omixLitMiner}
\alias{omixLitMiner}
\title{Generates a list of PubMed results based on a UniProt ID and a user-defined keyword.}
\usage{
omixLitMiner(df, output.file = NULL, plots.dir = NULL)
}
\arguments{
\item{df}{R dataframe having 5 attributes, UniProtID, IDType, TaxID, Keyword, KeywordInTitleOnly.}

\item{output.file}{Path string to the location to write the output file with the summarised results of the Pubmed query. Default = NULL gives no .xls summary file written out.}

\item{plots.dir}{Path to the directory where the plots generated by the package will be saved.Default = NULL gives no plots saved to disk.}
}
\value{
Generates an Excel file \code{output.file} with Pubmed query results using the UniProt identifers and keyword search in \code{query.file}.
}
\description{
Generates a list of PubMed results based on a UniProt ID and a user-defined keyword.
}
