% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PubMedMiner.R
\name{pubmedMiner_entry}
\alias{pubmedMiner_entry}
\title{Generates a list of Pubmed query results.}
\usage{
pubmedMiner_entry(query.file, output.file = "pubmed_results.xlsx")
}
\arguments{
\item{query.file}{Path to the input  file having the UniProtID.}

\item{output.file}{Path to the output  file with the results of the Pubmed query.}
}
\value{
Generates an Excel file \code{output.file} with Pubmed query results using the UniProt identifers and keyword search in \code{query.file}.
}
\description{
Generates a list of Pubmed query results.
}
\examples{
pubmedMiner_entry(query.file="potential_marker.xlsx", output.file = "potential_marker_pubmed_results.xlsx")
}
