% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PubMedMiner.R
\name{pubmedMiner_entry}
\alias{pubmedMiner_entry}
\title{Generates a list of PubMed results based on a UniProt ID and a user-defined keyword.}
\usage{
pubmedMiner_entry(df, output.file = "pubmed_results.xlsx",
  plots.dir = NULL)
}
\arguments{
\item{df}{R dataframe having 5 attributes, UniProtID, IDType, TaxID, Keyword, KeywordInTitleOnly.}

\item{output.file}{Path to the output file with the results of the PubMed query.}

\item{plots.dir}{Path to the directory where the plots generated by the package will be saved.}
}
\value{
Generates an Excel file \code{output.file} with Pubmed query results using the UniProt identifers and keyword search in \code{query.file}.
}
\description{
Generates a list of PubMed results based on a UniProt ID and a user-defined keyword.
}
\examples{
pubmedMiner_entry(potentialmarker, output.file = "potential_marker_pubmed_results.xlsx", plots.dir = 'plots')
df <- openxlsx::readWorkbook(system.file("extdata", "input_uniprot_keywords.xlsx", package="proteomicsliteratureminer"))
pubmedMiner_entry(df, output.file = "potential_marker_pubmed_results.xlsx", , plots.dir = 'plots')
}
