% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PubMedMiner.R
\name{pubmedMiner_entry}
\alias{pubmedMiner_entry}
\title{Generates a list of Pubmed results based on a UniProt ID and a user-defined keyword.}
\usage{
pubmedMiner_entry(df, output.file = "pubmed_results.xlsx")
}
\arguments{
\item{df}{R dataframe having 5 attributes, UniProtID, IDType, TaxID, Keyword, KeywordInTitleOnly.}

\item{output.file}{Path to the output  file with the results of the Pubmed query.}
}
\value{
Generates an Excel file \code{output.file} with Pubmed query results using the UniProt identifers and keyword search in \code{query.file}.
}
\description{
Generates a list of Pubmed results based on a UniProt ID and a user-defined keyword.
}
\examples{
pubmedMiner_entry(potentialmarker, output.file = "potential_marker_pubmed_results.xlsx")
df <- openxlsx::readWorkbook("~/Input.xlsx")
pubmedMiner_entry(df, output.file = "~/Pubmed_results.xlsx")
}
